requirements.txt
Flask>=2.3.3,<3.0
yfinance==0.2.37
pandas>=2.1,<2.2
numpy>=1.26,<1.27
scikit-learn>=1.4,<1.5
tensorflow-cpu>=2.18,<2.19  # Use CPU version instead of full TensorFlow
gunicorn>=20.1
